+++ addons/yat/src/attributes/IgnoreAttribute.cs
@@ -1,0 +1,6 @@
+using System;
+
+namespace YAT.Attributes;
+
+[AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
+public sealed class IgnoreAttribute : Attribute { }